pyproject.toml
[tool.poetry]
name = "airflow-driven-ml-pipeline"
version = "0.1.0"
description = "Airflow-driven MLOps pipeline for IMDb sentiment analysis"
authors = ["Your Name <you@example.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
apache-airflow = {version = "^2.8", extras = ["docker"]}
apache-airflow-providers-amazon = "^8.24.0"
apache-airflow-providers-slack = "^7.7.1"
great-expectations = "^0.18.12"
mlflow = "^2.17.0"
minio = "^7.2.7"
pandas = "^2.2.2"
scikit-learn = "^1.5.2"
fastapi = "^0.115.0"
uvicorn = {version = "^0.30.6", extras = ["standard"]}
evidently = "^0.4.20"
prometheus-client = "^0.21.0"
datasets = "^3.0.1"
sqlalchemy = "^2.0.31"
psycopg2-binary = "^2.9.9"
boto3 = "^1.35.26"
pyyaml = "^6.0.2"
hydra-core = "^1.3.2"
matplotlib = "^3.9.2"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
pytest-mock = "^3.14.0"
types-requests = "^2.32.0.20240914"
mypy = "^1.11.2"
ruff = "^0.7.0"
black = "^24.8.0"
pre-commit = "^3.8.0"

[tool.poetry.scripts]
seed-imdb = "scripts.seed_data:main"

[tool.mypy]
python_version = "3.11"
packages = ["src"]
strict = true
warn_unused_configs = true
warn_return_any = true
warn_unused_ignores = true

[tool.ruff]
target-version = "py311"
line-length = 100
src = ["src", "dags", "tests"]

[tool.ruff.lint]
select = ["E", "F", "I"]


